requirements.txt
discord.py
flask
